fix: Fixed a regression that caused the redirect loop detection to no longer function ([#348](https://github.com/nystudio107/craft-retour/issues/348))

+++ src/services/Redirects.php
@@ -1086,9 +1086,10 @@ class Redirects extends Component
 
     /**
      * @param array $redirectConfig
-     * @return bool whether the redirect was saved or not
+     * @param bool $checkForRedirectLoop
+     * @return bool
      */
-    public function saveRedirect(array $redirectConfig): bool
+    public function saveRedirect(array $redirectConfig, bool $checkForRedirectLoop = true): bool
     {
         // Handle URL encoded URLs by decoding them before saving them
         if (isset($redirectConfig['redirectMatchType']) && $redirectConfig['redirectMatchType'] === 'exactmatch') {
@@ -1201,28 +1202,30 @@ class Redirects extends Component
                 return false;
             }
         }
-        // To prevent redirect loops, see if any static redirects have our redirectDestUrl as their redirectSrcUrl
-        $testRedirectConfig = $this->getRedirectByRedirectSrcUrl(
-            $redirectConfig['redirectDestUrl'],
-            $redirectConfig['siteId']
-        );
-        if ($testRedirectConfig !== null) {
-            Craft::debug(
-                Craft::t(
-                    'retour',
-                    'Deleting redirect to prevent a loop: {redirect}',
-                    ['redirect' => print_r($testRedirectConfig, true)]
-                ),
-                __METHOD__
+        if ($checkForRedirectLoop) {
+            // To prevent redirect loops, see if any static redirects have our redirectDestUrl as their redirectSrcUrl
+            $testRedirectConfig = $this->getRedirectByRedirectSrcUrl(
+                $redirectConfig['redirectDestUrl'],
+                $redirectConfig['siteId']
             );
-            // Delete the redirect that has a redirectSrcUrl the same as this record's redirectDestUrl
-            try {
-                $db->createCommand()->delete(
-                    '{{%retour_static_redirects}}',
-                    ['id' => $testRedirectConfig['id']]
-                )->execute();
-            } catch (Exception $e) {
-                Craft::error($e->getMessage(), __METHOD__);
+            if ($testRedirectConfig !== null) {
+                Craft::debug(
+                    Craft::t(
+                        'retour',
+                        'Deleting redirect to prevent a loop: {redirect}',
+                        ['redirect' => print_r($testRedirectConfig, true)]
+                    ),
+                    __METHOD__
+                );
+                // Delete the redirect that has a redirectSrcUrl the same as this record's redirectDestUrl
+                try {
+                    $db->createCommand()->delete(
+                        '{{%retour_static_redirects}}',
+                        ['id' => $testRedirectConfig['id']]
+                    )->execute();
+                } catch (Exception $e) {
+                    Craft::error($e->getMessage(), __METHOD__);
+                }
             }
         }
         // Trigger a 'afterSaveRedirect' event
